refactor(telemetry): share the once-per-page-load failure claim set
The session-ingest and daemon-relay senders each held their own claim set
for once-per-condition Sentry reporting, the second copy of one behavior.
Both now claim through lib/telemetry/report-once.ts, with namespaced
condition strings so the senders cannot claim each other's, and both test
resets clear the one set.
